requirements.txt
Box2D==2.3.2
Box2D-kengz==2.3.3
gym==0.14.0
matplotlib==3.1.1
numpy==1.17.2
tensorflow-gpu==2.0.0b1
